pyproject.toml
[project]
name = "my-blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-ckeditor-5>=0.2.15",
    "django-imagekit>=5.0.0",
    "django>=5.1.3",
    "pillow>=11.0.0",
]
